[Refactor] Refactor the tools source file directory
